requirements.txt
asgiref==3.7.2
Django==4.2.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
flake8-quotes==3.3.2
flake8-variables-names==0.0.6
mccabe==0.7.0
pep8-naming==0.13.3
Pillow==9.5.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
